+++ src/constants.js
@@ -40,10 +40,18 @@ const State = {
   ABORTING: 'ABORTING'
 };
 
+const FILTER_TYPES = {
+  ALL: 'all-time',
+  TODAY: 'today',
+  WEEK: 'week',
+  MONTH: 'month',
+  YEAR: 'year',
+};
+
 const ENTER_KEYCODE = 10;
 
 const ENTER_ALT_KEYCODE = 13;
 
 const MAX_SHORT_DESCRIPTION_LENGTH = 140;
 
-export {State, POPUP_BUTTON_ACTIVE_CLASS_NAME, CARD_BUTTON_ACTIVE_CLASS_NAME, SortType, MAX_SHORT_DESCRIPTION_LENGTH, UpdateType, UserAction, FilterType, ENTER_KEYCODE, ENTER_ALT_KEYCODE, MenuItem};
+export {State, POPUP_BUTTON_ACTIVE_CLASS_NAME, CARD_BUTTON_ACTIVE_CLASS_NAME, SortType, MAX_SHORT_DESCRIPTION_LENGTH, UpdateType, UserAction, FilterType, ENTER_KEYCODE, ENTER_ALT_KEYCODE, MenuItem, FILTER_TYPES};
