+++ src/view/loading-template.js
@@ -1,0 +1,1 @@
+export const createLoadingTemplate = () => '<h2 class="films-list__title">Loading...</h2>';
