+++ src/view/loading-view.js
@@ -1,6 +1,5 @@
 import AbstractView from './abstract-view';
-
-const createLoadingTemplate = () => '<h2 class="films-list__title">Loading...</h2>';
+import {createLoadingTemplate} from './loading-template';
 
 export default class LoadingView extends AbstractView {
   get template() {
