+++ src/view/movie-list-view.js
@@ -1,8 +1,5 @@
 import AbstractView from './abstract-view';
-
-const createMovieListTemplate = () => `<section class="films-list">
-      <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
-    </section>`;
+import {createMovieListTemplate} from './templates/movie-list-template';
 
 export default class MovieListView extends AbstractView {
   get template() {
