+++ src/view/movies-section-view.js
@@ -1,6 +1,5 @@
 import AbstractView from './abstract-view';
-
-const createMovieSectionTemplate = () => '<section class="films"></section>';
+import {createMovieSectionTemplate} from './templates/movies-section-template';
 
 export default class MoviesSectionView extends AbstractView {
   get template() {
