+++ src/view/templates/load-more-button-template.js
@@ -1,0 +1,3 @@
+const createLoadMoreButtonTemplate = () => '<button class="films-list__show-more">Show more</button>';
+
+export {createLoadMoreButtonTemplate};
