+++ src/view/templates/movie-count-template.js
@@ -1,0 +1,3 @@
+const createMoviesCountTemplate = (moviesCount) => `<p>${moviesCount} movies inside</p>`;
+
+export {createMoviesCountTemplate};
