+++ src/view/templates/movie-list-container-template.js
@@ -1,0 +1,3 @@
+const createMovieListContainerTemplate = () => '<div class="films-list__container"></div>';
+
+export {createMovieListContainerTemplate};
