+++ src/view/templates/movie-list-template.js
@@ -1,0 +1,5 @@
+const createMovieListTemplate = () => `<section class="films-list">
+      <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
+    </section>`;
+
+export {createMovieListTemplate};
