+++ src/view/templates/movies-section-template.js
@@ -1,0 +1,3 @@
+const createMovieSectionTemplate = () => '<section class="films"></section>';
+
+export {createMovieSectionTemplate};
