+++ src/view/templates/site-menu-template.js
@@ -1,0 +1,5 @@
+const createSiteMenuTemplate = () => (
+  '<nav class="main-navigation"></nav>'
+);
+
+export {createSiteMenuTemplate};
