+++ src/view/templates/stats-menu-item-template.js
@@ -1,0 +1,3 @@
+const createStatsMenuItemTemplate = (isActive) => `<a href="#stats" class="main-navigation__additional${isActive ? ' main-navigation__additional--active' : ''}">Stats</a>`;
+
+export {createStatsMenuItemTemplate};
